Move folder from magnets to magnet
