+++ src/dodal/devices/beamlines/i06_1/magnet/coordinates.py
