+++ src/dodal/devices/beamlines/i06_1/magnet/enums.py
