+++ tests/devices/beamlines/i06_1/magnet/utils.py
@@ -1,6 +1,6 @@
 import math
 
-from dodal.devices.beamlines.i06_1.magnets.coordinates import (
+from dodal.devices.beamlines.i06_1.magnet.coordinates import (
     MagnetPosition,
     MagnetSphericalPosition,
 )
